begin refactoring/reorganizing
